+++ apps/server/src/modules/common/common.definition.ts
@@ -1,0 +1,6 @@
+import { CxModule } from "@cxnext/core";
+
+export class CommonDefinition extends CxModule {
+  public readonly name = "common";
+  public readonly boundedContext = "common";
+}