+++ apps/server/src/modules/common/common.registry.ts
@@ -1,0 +1,15 @@
+import { Inject, Injectable, type OnModuleInit } from "@nestjs/common";
+import { ModuleRegistryService } from "../../core/modules/module-registry.service";
+import { CommonDefinition } from "./common.definition";
+
+@Injectable()
+export class CommonRegistryBootstrap implements OnModuleInit {
+  public constructor(
+    @Inject(ModuleRegistryService)
+    private readonly moduleRegistryService: ModuleRegistryService,
+  ) {}
+
+  public onModuleInit(): void {
+    this.moduleRegistryService.register(new CommonDefinition());
+  }
+}