+++ apps/server/src/modules/common/database/migrations/001-create-common-location.ts
@@ -1,0 +1,1 @@
+export { createCommonLocationMigration } from "@cxnext/db";