+++ apps/server/src/modules/common/database/migrations/common-master-migrations.ts
@@ -1,0 +1,7 @@
+export {
+  contactsCommonMigrations,
+  locationCommonMigrations,
+  ordersCommonMigrations,
+  othersCommonMigrations,
+  productCommonMigrations,
+} from "@cxnext/db";