+++ apps/server/src/modules/common/database/migrations/index.ts
@@ -1,0 +1,8 @@
+export {
+  contactsCommonMigrations,
+  createCommonLocationMigration,
+  locationCommonMigrations,
+  ordersCommonMigrations,
+  othersCommonMigrations,
+  productCommonMigrations,
+} from "@cxnext/db";