+++ apps/server/src/modules/common/database/seeder/common-master-seeders.ts
@@ -1,0 +1,7 @@
+export {
+  contactsCommonSeeders,
+  locationCommonSeeders,
+  ordersCommonSeeders,
+  othersCommonSeeders,
+  productCommonSeeders,
+} from "@cxnext/db";