+++ apps/server/src/modules/common/database/seeder/index.ts
@@ -1,0 +1,8 @@
+export {
+  contactsCommonSeeders,
+  locationCommonSeeders,
+  ordersCommonSeeders,
+  othersCommonSeeders,
+  productCommonSeeders,
+  seedCommonLocationSeeder,
+} from "@cxnext/db";